pyproject.toml
[project]
name = "cryptoassetsmanager"
version = "0.1.0"
description = "Crypto Assets Manager"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.4",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.6.1",
    "ccxt>=4.4.35",
    "pandas>=2.2.3",
    "motor>=3.6.0",
    "pymongo>=4.9.2",
    "apscheduler>=3.11.0"
]

[tool.ruff]
line-length = 88
target-version = "py312"

exclude = [
  ".git",
  ".ruff_cache", 
  "venv",
  "__pycache__",
]

[tool.ruff.lint]
select = [
   "E",   # pycodestyle
   "W",   # pycodestyle warnings
   "F",   # pyflakes
]
ignore = ["E501", "W292", "W291", "W293"]
per-file-ignores = { "__init__.py" = ["F401"] }

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"